finance

+++ src/main/java/com/yimei/finance/repository/admin/applyinfo/ApplyInfoRepository.java
@@ -6,5 +6,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by zhangbolun on 16/8/16.
  */
-public interface ApplyInfoRepository extends JpaRepository<ApplyInfo, Long> {
+public interface ApplyInfoRepository extends JpaRepository<ApplyInfo, Integer> {
+
+//    @Modifying(clearAutomatically = true)
+//    @Query(" update ApplyInfo applyInfo set applyInfo.applyType=:applyType where applyInfo.id=:id")
+//    void updateApplyInfoType(@Param("id") Integer id,
+//                             @Param("applyType") String applyType);
 }
